implementada la clase Desgaste e incorporada la funcionalidad en Hacha

+++ src/fiuba/algo3/modelo/Desgaste.java
@@ -2,25 +2,25 @@ package fiuba.algo3.modelo;
 
 public class Desgaste {
 
-    private int durabilidad;
+    private double durabilidad;
     private int fuerza;
     private int factorDesgaste;
 
-    public Desgaste(int durabilidadInicial, int fuerza){
+    public Desgaste(double durabilidadInicial, int fuerza){
 
         this.durabilidad = durabilidadInicial;
         this.fuerza = fuerza;
         this.factorDesgaste = 1;
     }
 
-    public Desgaste(int durabilidadInicial, int fuerza, int factorDesgaste){
+    public Desgaste(double durabilidadInicial, int fuerza, int factorDesgaste){
 
         this.durabilidad = durabilidadInicial;
         this.fuerza = fuerza;
         this.factorDesgaste = factorDesgaste;
     }
 
-    public int desgastar() {
+    public double desgastar() {
 
         durabilidad -= fuerza/factorDesgaste;
